Only allow document api protocol version 6.221 (Java)
